refactor: fix email service interfaces import
